refactor(#72): 장소 및 시간 투표 추상화
- 장소와 시간 투표 및 확정 실행객체 인터페이스로 추상화
- 나의 투표(장소, 시간) 조회 인터페이스로 추상화
