+++ src/main/java/com/prography/yakgwa/domain/meet/controller/MeetCrateController.java
@@ -1,6 +1,7 @@
 package com.prography.yakgwa.domain.meet.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.prography.yakgwa.domain.meet.controller.docs.MeetCreateApi;
 import com.prography.yakgwa.domain.meet.controller.req.CreateMeetRequest;
 import com.prography.yakgwa.domain.meet.controller.res.CreateMeetResponse;
 import com.prography.yakgwa.domain.meet.entity.Meet;
