+++ src/main/java/com/prography/yakgwa/domain/meet/entity/MeetStatus.java
@@ -1,5 +1,13 @@
 package com.prography.yakgwa.domain.meet.entity;
 
 public enum MeetStatus {
-    CONFIRM,BEFORE_CONFIRM,BEFORE_VOTE,VOTE
+    CONFIRM, BEFORE_CONFIRM, BEFORE_VOTE, VOTE;
+
+    public boolean isConfirm() {
+        return this == CONFIRM;
+    }
+
+    public boolean isBeforeConfirm() {
+        return this == BEFORE_CONFIRM;
+    }
 }
