+++ src/main/java/com/prography/yakgwa/domain/meet/impl/MeetStatusJudger.java
@@ -9,7 +9,7 @@ import com.prography.yakgwa.domain.vote.repository.PlaceSlotJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.PlaceVoteJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.TimeSlotJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.TimeVoteJpaRepository;
-import com.prography.yakgwa.domain.vote.service.VoteCounter;
+import com.prography.yakgwa.domain.vote.service.impl.VoteCounter;
 import com.prography.yakgwa.global.meta.ImplService;
 import lombok.RequiredArgsConstructor;
