+++ src/main/java/com/prography/yakgwa/domain/meet/service/MeetService.java
@@ -1,28 +1,19 @@
 package com.prography.yakgwa.domain.meet.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.prography.yakgwa.domain.common.schedule.AlarmScheduler;
 import com.prography.yakgwa.domain.meet.entity.Meet;
 import com.prography.yakgwa.domain.meet.entity.MeetStatus;
 import com.prography.yakgwa.domain.meet.impl.MeetStatusJudger;
-import com.prography.yakgwa.domain.meet.impl.MeetWriter;
 import com.prography.yakgwa.domain.meet.repository.MeetJpaRepository;
-import com.prography.yakgwa.domain.meet.service.req.MeetCreateRequestDto;
 import com.prography.yakgwa.domain.meet.service.req.MeetWithVoteAndStatus;
 import com.prography.yakgwa.domain.meet.service.res.MeetInfoWithParticipant;
 import com.prography.yakgwa.domain.participant.entity.Participant;
-import com.prography.yakgwa.domain.participant.impl.ParticipantWriter;
 import com.prography.yakgwa.domain.participant.repository.ParticipantJpaRepository;
-import com.prography.yakgwa.domain.place.entity.Place;
-import com.prography.yakgwa.domain.place.entity.dto.PlaceInfoDto;
-import com.prography.yakgwa.domain.place.repository.PlaceJpaRepository;
 import com.prography.yakgwa.domain.user.entity.User;
 import com.prography.yakgwa.domain.user.repository.UserJpaRepository;
 import com.prography.yakgwa.domain.vote.entity.place.PlaceSlot;
 import com.prography.yakgwa.domain.vote.entity.time.TimeSlot;
 import com.prography.yakgwa.domain.vote.repository.PlaceSlotJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.TimeSlotJpaRepository;
-import com.prography.yakgwa.global.format.exception.meet.ConfirmPlaceCountException;
 import com.prography.yakgwa.global.format.exception.meet.NotFoundMeetException;
 import com.prography.yakgwa.global.format.exception.param.DataIntegrateException;
 import com.prography.yakgwa.global.format.exception.user.NotFoundUserException;
@@ -72,11 +63,31 @@ public class MeetService {
         User user = userJpaRepository.findById(userId)
                 .orElseThrow(NotFoundUserException::new);
         List<Participant> participants = participantJpaRepository.findAllByUserId(userId);
-        return participants.stream()
-                .map(participant -> createMeetWithVoteAndStatus(participant, user))
-                .filter(meet -> meet.getMeetStatus().equals(MeetStatus.CONFIRM) ||
-                        meet.getMeetStatus().equals(MeetStatus.BEFORE_CONFIRM))
-                .toList();
+
+        List<MeetWithVoteAndStatus> meetsForUser = new ArrayList<>();
+        for (Participant participant : participants) {
+            Meet meet = participant.getMeet();
+            MeetStatus meetStatus = meetStatusJudger.judge(meet, user);
+            if (isMeetStatusAfterVote(meetStatus)) {
+                PlaceSlot placeSlot = getConfirmPlaceSlot(meet);
+                TimeSlot timeSlot = getConfirmTimeSlot(meet);
+                meetsForUser.add(MeetWithVoteAndStatus.of(meet, timeSlot, placeSlot, meetStatus));
+            }
+        }
+        return meetsForUser;
+    }
+
+    private static boolean isMeetStatusAfterVote(MeetStatus meetStatus) {
+        return meetStatus.isConfirm() || meetStatus.isBeforeConfirm();
+    }
+
+    private MeetWithVoteAndStatus createMeetWithVoteAndStatus(Participant participant, User user) {
+        Meet meet = participant.getMeet();
+
+        MeetStatus meetStatus = meetStatusJudger.judge(meet, user);
+        PlaceSlot placeSlot = getConfirmPlaceSlot(meet);
+        TimeSlot timeSlot = getConfirmTimeSlot(meet);
+        return MeetWithVoteAndStatus.of(meet, timeSlot, placeSlot, meetStatus);
     }
 
     /**
@@ -103,13 +114,6 @@ public class MeetService {
         return list;
     }
 
-    private MeetWithVoteAndStatus createMeetWithVoteAndStatus(Participant participant, User user) {
-        Meet meet = participant.getMeet();
-        MeetStatus meetStatus = meetStatusJudger.judge(meet, user);
-        PlaceSlot placeSlot = getConfirmPlaceSlot(meet);
-        TimeSlot timeSlot = getConfirmTimeSlot(meet);
-        return MeetWithVoteAndStatus.of(meet, timeSlot, placeSlot, meetStatus);
-    }
 
     private boolean shouldSkipMeet(MeetWithVoteAndStatus meet) {
         LocalDateTime now = LocalDateTime.now();
