+++ src/main/java/com/prography/yakgwa/domain/place/controller/docs/PlaceApi.java
@@ -1,4 +1,4 @@
-package com.prography.yakgwa.domain.place.controller;
+package com.prography.yakgwa.domain.place.controller.docs;
 
 import com.prography.yakgwa.domain.place.controller.req.LikePlaceRequest;
 import com.prography.yakgwa.domain.place.service.dto.PlaceInfoWithUserLike;
