+++ src/main/java/com/prography/yakgwa/domain/vote/service/VoteFinder.java
@@ -1,0 +1,5 @@
+package com.prography.yakgwa.domain.vote.service;
+
+public interface VoteFinder<T> {
+    T findVoteInfoWithStatusOf(Long userId, Long meetId);
+}
