+++ src/main/java/com/prography/yakgwa/domain/vote/service/time/TimeVoteExecuteService.java
@@ -1,25 +1,25 @@
-package com.prography.yakgwa.domain.vote.service;
+package com.prography.yakgwa.domain.vote.service.time;
 
+import com.prography.yakgwa.domain.common.schedule.AlarmScheduler;
 import com.prography.yakgwa.domain.meet.entity.Meet;
+import com.prography.yakgwa.domain.meet.impl.MeetStatusJudger;
 import com.prography.yakgwa.domain.meet.repository.MeetJpaRepository;
+import com.prography.yakgwa.domain.participant.entity.Participant;
 import com.prography.yakgwa.domain.participant.repository.ParticipantJpaRepository;
 import com.prography.yakgwa.domain.user.entity.User;
 import com.prography.yakgwa.domain.user.repository.UserJpaRepository;
-import com.prography.yakgwa.domain.vote.entity.place.PlaceSlot;
-import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
 import com.prography.yakgwa.domain.vote.entity.time.TimeSlot;
 import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
-import com.prography.yakgwa.domain.vote.repository.PlaceSlotJpaRepository;
-import com.prography.yakgwa.domain.vote.repository.PlaceVoteJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.TimeSlotJpaRepository;
 import com.prography.yakgwa.domain.vote.repository.TimeVoteJpaRepository;
+import com.prography.yakgwa.domain.vote.service.VoteExecuter;
 import com.prography.yakgwa.domain.vote.service.req.EnableTimeRequestDto;
 import com.prography.yakgwa.global.format.exception.meet.NotFoundMeetException;
 import com.prography.yakgwa.global.format.exception.participant.NotFoundParticipantException;
+import com.prography.yakgwa.global.format.exception.slot.NotFoundTimeSlotException;
 import com.prography.yakgwa.global.format.exception.user.NotFoundUserException;
-import com.prography.yakgwa.global.format.exception.vote.AlreadyPlaceConfirmException;
 import com.prography.yakgwa.global.format.exception.vote.AlreadyTimeConfirmVoteException;
-import com.prography.yakgwa.global.format.exception.vote.NotValidVotePlaceException;
+import com.prography.yakgwa.global.format.exception.vote.NotValidConfirmTimeException;
 import com.prography.yakgwa.global.format.exception.vote.NotValidVoteTimeException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -27,39 +27,28 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Set;
 
 @Transactional
-@RequiredArgsConstructor
 @Service
-public class VoteExecuteService {
-    private final PlaceSlotJpaRepository placeSlotJpaRepository;
-    private final UserJpaRepository userJpaRepository;
-    private final PlaceVoteJpaRepository placeVoteJpaRepository;
-    private final TimeSlotJpaRepository timeSlotJpaRepository;
-    private final TimeVoteJpaRepository timeVoteJpaRepository;
+@RequiredArgsConstructor
+public class TimeVoteExecuteService implements VoteExecuter<TimeVote, EnableTimeRequestDto> {
     private final MeetJpaRepository meetJpaRepository;
     private final ParticipantJpaRepository participantJpaRepository;
+    private final TimeSlotJpaRepository timeSlotJpaRepository;
+    private final MeetStatusJudger meetStatusJudger;
+    private final AlarmScheduler alarmScheduler;
+    private final UserJpaRepository userJpaRepository;
+    private final TimeVoteJpaRepository timeVoteJpaRepository;
 
-    /**
-     * 1. 투표확정여부 확인
-     * 2. 기존의 투표목록 삭제
-     * 3. 시간 투표항목에 없는 시간이 들어왔는지 확인 -> 새로운 항목이라면 추가
-     * 4. 시간 투표 생성
-     */
-    /**
-     * Work) Test Code
-     * Write-Date) 2024-07-29, 월, 14:21
-     * Finish-Date) 2024-07-30
-     */
-    public List<TimeVote> voteTime(Long userId, Long meetId, EnableTimeRequestDto requestDto) {
-        if (!participantJpaRepository.existsByUserIdAndMeetId(userId, meetId)) {
+    @Override
+    public List<TimeVote> vote(Long userId, Long meetId, EnableTimeRequestDto requestDto) {
+        if (isNotExistParticipant(userId, meetId)) {
             throw new NotFoundParticipantException();
         }
         if (isTimeConfirmedFrom(meetId)) {
             throw new AlreadyTimeConfirmVoteException();
         }
-        Meet meet = findBy(meetId);
+        Meet meet = findMeetById(meetId);
         List<TimeSlot> allTimeSlotsInMeet = timeSlotJpaRepository.findAllByMeetId(meetId);
         if (isValidTimeVote(allTimeSlotsInMeet, meet)) {
             throw new AlreadyTimeConfirmVoteException();
@@ -89,48 +78,50 @@ public class VoteExecuteService {
                 .timeSlot(timeSlot)
                 .user(user)
                 .build()).toList();
-        List<TimeVote> timeVotes1 = timeVoteJpaRepository.saveAll(timeVotes);
-        return timeVotes1;
+        return timeVoteJpaRepository.saveAll(timeVotes);
+
     }
 
+
     /**
-     * 1. 확정여부 확인
-     * 2. 기존투표 삭제
-     * 3. 새로운 투표생성
+     * Work) 테스트코드
+     * Write-Date) 2024-07-25
+     * Finish-Date) 2024-07-30
      */
     /**
-     * Work) Test Code
-     * Write-Date) 2024-07-29, 월, 14:21
+     * Todo
+     * Work) 확정가능시간에대한 테스트코드
+     * Write-Date)
      * Finish-Date)
      */
-    public List<PlaceVote> votePlace(Long userId, Long meetId, Set<Long> placeSlotIds) {
-        if (placeSlotJpaRepository.isConfirmFrom(meetId)) {
-            throw new AlreadyPlaceConfirmException();
+    @Override
+    public void confirm(Long meetId, Long userId, Long timeSlotId) {
+        Meet meet = findMeetById(meetId);
+        if (isTimeConfirmedFrom(meetId)) {
+            throw new AlreadyTimeConfirmVoteException();
         }
-        if (!participantJpaRepository.existsByUserIdAndMeetId(userId, meetId)) {
-            throw new NotFoundParticipantException();
+        if (meet.getValidConfirmTime().isBefore(LocalDateTime.now())) {
+            throw new NotValidConfirmTimeException();
         }
-        List<PlaceSlot> placeSlots = placeSlotJpaRepository.findAllByMeetId(meetId);
-        boolean isContainSlot = placeSlots.stream().allMatch(placeSlot -> placeSlotIds.contains(placeSlot.getId()));
-        if (!isContainSlot) {
-            throw new NotValidVotePlaceException();
+        Participant participant = participantJpaRepository.findByUserIdAndMeetId(userId, meetId)
+                .orElseThrow(NotFoundParticipantException::new);
+        TimeSlot timeSlot = timeSlotJpaRepository.findById(timeSlotId)
+                .orElseThrow(NotFoundTimeSlotException::new);
+        timeSlot.confirm();
+        if (meetStatusJudger.isConfirm(meet)) {
+            alarmScheduler.registerAlarm(meet);
         }
-        User user = userJpaRepository.findById(userId).orElseThrow(NotFoundUserException::new);
-        placeVoteJpaRepository.deleteAllByUserIdAndMeetId(user, meetId);
-        List<PlaceSlot> choosePlaceSlot = placeSlotJpaRepository.findAllById(placeSlotIds);
-        List<PlaceVote> placeVotes = choosePlaceSlot.stream()
-                .map(placeSlot -> PlaceVote.builder()
-                        .placeSlot(placeSlot)
-                        .user(user)
-                        .build())
-                .toList();
-        return placeVoteJpaRepository.saveAll(placeVotes);
     }
 
-    private Meet findBy(Long meetId) {
-        return meetJpaRepository.findById(meetId)
-                .orElseThrow(NotFoundMeetException::new);
+
+    private boolean isNotExistParticipant(Long userId, Long meetId) {
+        return !participantJpaRepository.existsByUserIdAndMeetId(userId, meetId);
     }
+
+    private Meet findMeetById(Long meetId) {
+        return meetJpaRepository.findById(meetId).orElseThrow(NotFoundMeetException::new);
+    }
+
 
     private boolean isTimeConfirmedFrom(Long meetId) {
         return timeSlotJpaRepository.existsConfirmFrom(meetId);
