+++ src/test/java/com/prography/yakgwa/domain/vote/service/VoteExecuteServiceTest.java
@@ -5,21 +5,17 @@ import com.prography.yakgwa.testHelper.RepositoryDeleter;
 import com.prography.yakgwa.domain.meet.entity.Meet;
 import com.prography.yakgwa.domain.meet.entity.MeetTheme;
 import com.prography.yakgwa.domain.participant.entity.enumerate.MeetRole;
-import com.prography.yakgwa.domain.place.entity.Place;
 import com.prography.yakgwa.domain.user.entity.User;
-import com.prography.yakgwa.domain.vote.entity.place.PlaceSlot;
-import com.prography.yakgwa.domain.vote.entity.place.PlaceVote;
 import com.prography.yakgwa.domain.vote.entity.time.TimeSlot;
 import com.prography.yakgwa.domain.vote.entity.time.TimeVote;
 import com.prography.yakgwa.domain.vote.repository.TimeSlotJpaRepository;
 import com.prography.yakgwa.domain.vote.service.req.EnableTimeRequestDto;
 import com.prography.yakgwa.global.format.exception.participant.NotFoundParticipantException;
-import com.prography.yakgwa.global.format.exception.vote.AlreadyPlaceConfirmException;
-import com.prography.yakgwa.global.format.exception.vote.NotValidVotePlaceException;
 import com.prography.yakgwa.global.format.exception.vote.NotValidVoteTimeException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -33,12 +29,13 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @ActiveProfiles("test")
 @SpringBootTest
-class VoteExecuteServiceTest {
+class TimeSlotExecuteServiceTest {
     @Autowired
     DummyCreater dummyCreater;
 
     @Autowired
-    VoteExecuteService voteExecuteService;
+    @Qualifier("timeVoteExecuteService")
+    VoteExecuter<TimeVote, EnableTimeRequestDto> voteExecuter;
     @Autowired
     private TimeSlotJpaRepository timeSlotJpaRepository;
     @Autowired
@@ -72,7 +69,7 @@ class VoteExecuteServiceTest {
         // when
         System.out.println("=====Logic Start=====");
 
-        List<TimeVote> timeVotes = voteExecuteService.voteTime(saveUser.getId(), saveMeet.getId(), request);
+        List<TimeVote> timeVotes = voteExecuter.vote(saveUser.getId(), saveMeet.getId(), request);
 
         System.out.println("=====Logic End=====");
         // then
@@ -106,7 +103,7 @@ class VoteExecuteServiceTest {
         // when
         System.out.println("=====Logic Start=====");
 
-        List<TimeVote> timeVotes = voteExecuteService.voteTime(saveUser.getId(), saveMeet.getId(), request);
+        List<TimeVote> timeVotes = voteExecuter.vote(saveUser.getId(), saveMeet.getId(), request);
 
         System.out.println("=====Logic End=====");
         // then
@@ -136,7 +133,7 @@ class VoteExecuteServiceTest {
         // when
         System.out.println("=====Logic Start=====");
 
-        List<TimeVote> timeVotes = voteExecuteService.voteTime(saveUser.getId(), saveMeet.getId(), request);
+        List<TimeVote> timeVotes = voteExecuter.vote(saveUser.getId(), saveMeet.getId(), request);
 
         System.out.println("=====Logic End=====");
         // then
@@ -165,7 +162,7 @@ class VoteExecuteServiceTest {
         // when
         // then
         System.out.println("=====Logic Start=====");
-        assertThrows(NotFoundParticipantException.class, () -> voteExecuteService.voteTime(saveUser.getId(), saveMeet.getId(), request));
+        assertThrows(NotFoundParticipantException.class, () -> voteExecuter.vote(saveUser.getId(), saveMeet.getId(), request));
     }
 
     @Test
@@ -187,80 +184,8 @@ class VoteExecuteServiceTest {
         // when
         // then
         System.out.println("=====Logic Start=====");
-        assertThrows(NotValidVoteTimeException.class, () -> voteExecuteService.voteTime(saveUser.getId(), saveMeet.getId(), request));
+        assertThrows(NotValidVoteTimeException.class, () -> voteExecuter.vote(saveUser.getId(), saveMeet.getId(), request));
     }
 
-    /*================votePlace===============*/
-    @Test
-    void 이미장소가확정되어있는경우장소투표_예외() {
-        // given
-        User saveUser = dummyCreater.createAndSaveUser(1);
-        MeetTheme saveMeetTheme = dummyCreater.createAndSaveMeetTheme(1);
-        Meet saveMeet = dummyCreater.createAndSaveMeet(1, saveMeetTheme, 24);
-        dummyCreater.createAndSaveParticipant(saveMeet, saveUser, MeetRole.LEADER);
-        Place savePlace = dummyCreater.createAndSavePlace(1);
-        dummyCreater.createAndSavePlaceSlot(savePlace, saveMeet, true);
-        // when
-        // then
-        System.out.println("=====Logic Start=====");
-        assertThrows(AlreadyPlaceConfirmException.class, () -> voteExecuteService.votePlace(saveUser.getId(), saveMeet.getId(), Set.of(1L)));
-        System.out.println("=====Logic End=====");
-    }
 
-    @Test
-    void 추가가안되어있는장소후보의장소투표_예외() {
-        // given
-        User saveUser = dummyCreater.createAndSaveUser(1);
-        MeetTheme saveMeetTheme = dummyCreater.createAndSaveMeetTheme(1);
-        Meet saveMeet = dummyCreater.createAndSaveMeet(1, saveMeetTheme, 24);
-        dummyCreater.createAndSaveParticipant(saveMeet, saveUser, MeetRole.LEADER);
-        Place savePlace = dummyCreater.createAndSavePlace(1);
-        PlaceSlot savePlaceSlot = dummyCreater.createAndSavePlaceSlot(savePlace, saveMeet, false);
-        // when
-        // then
-        System.out.println("=====Logic Start=====");
-        assertThrows(NotValidVotePlaceException.class, () -> voteExecuteService.votePlace(saveUser.getId(), saveMeet.getId(), Set.of(savePlaceSlot.getId() + 1L)));
-        System.out.println("=====Logic End=====");
-    }
-
-    @Test
-    void 장소투표() {
-        // given
-        User saveUser = dummyCreater.createAndSaveUser(1);
-        MeetTheme saveMeetTheme = dummyCreater.createAndSaveMeetTheme(1);
-        Meet saveMeet = dummyCreater.createAndSaveMeet(1, saveMeetTheme, 24);
-        dummyCreater.createAndSaveParticipant(saveMeet, saveUser, MeetRole.LEADER);
-
-        Place savePlace1 = dummyCreater.createAndSavePlace(1);
-        Place savePlace2 = dummyCreater.createAndSavePlace(2);
-        PlaceSlot savePlaceSlot1 = dummyCreater.createAndSavePlaceSlot(savePlace1, saveMeet, false);
-        PlaceSlot savePlaceSlot2 = dummyCreater.createAndSavePlaceSlot(savePlace2, saveMeet, false);
-        // when
-        System.out.println("=====Logic Start=====");
-
-        List<PlaceVote> placeVotes = voteExecuteService.votePlace(saveUser.getId(), saveMeet.getId(), Set.of(savePlaceSlot1.getId(), savePlaceSlot2.getId()));
-
-        System.out.println("=====Logic End=====");
-        // then
-        assertAll(() -> assertThat(placeVotes.size()).isEqualTo(2));
-
-    }
-
-    @Test
-    void 모임에참여중인아닌데장소투표_예외() {
-        // given
-        User saveUser = dummyCreater.createAndSaveUser(1);
-        MeetTheme saveMeetTheme = dummyCreater.createAndSaveMeetTheme(1);
-        Meet saveMeet = dummyCreater.createAndSaveMeet(1, saveMeetTheme, 24);
-
-        Place savePlace1 = dummyCreater.createAndSavePlace(1);
-        Place savePlace2 = dummyCreater.createAndSavePlace(2);
-        PlaceSlot savePlaceSlot1 = dummyCreater.createAndSavePlaceSlot(savePlace1, saveMeet, false);
-        PlaceSlot savePlaceSlot2 = dummyCreater.createAndSavePlaceSlot(savePlace2, saveMeet, false);
-        // when
-        // then
-        System.out.println("=====Logic Start=====");
-        assertThrows(NotFoundParticipantException.class, () -> voteExecuteService.votePlace(saveUser.getId(), saveMeet.getId(), Set.of(savePlaceSlot1.getId(), savePlaceSlot2.getId())));
-        System.out.println("=====Logic End=====");
-    }
 }
